moved all core files to their own directory open_mpic_core
