+++ src/open_mpic_core/common_domain/check_request.py
@@ -1,7 +1,7 @@
 from abc import ABC
 from pydantic import BaseModel
 
-from aws_lambda_python.common_domain.check_parameters import CaaCheckParameters, DcvCheckParameters
+from open_mpic_core.common_domain.check_parameters import CaaCheckParameters, DcvCheckParameters
 
 
 class BaseCheckRequest(BaseModel, ABC):
