+++ src/open_mpic_core/common_domain/enum/check_type.py
