+++ src/open_mpic_core/common_domain/enum/dns_record_type.py
