+++ src/open_mpic_core/mpic_coordinator/domain/mpic_request.py
@@ -3,12 +3,11 @@ from typing import Literal, Union
 
 from typing_extensions import Annotated
 
-from aws_lambda_python.common_domain.enum.check_type import CheckType
+from open_mpic_core.common_domain.enum.check_type import CheckType
 from pydantic import BaseModel, model_validator, Field
 
-from aws_lambda_python.mpic_coordinator.domain.mpic_orchestration_parameters import MpicRequestOrchestrationParameters
-from aws_lambda_python.common_domain.check_parameters import CaaCheckParameters, DcvCheckParameters
-from aws_lambda_python.common_domain.enum.dcv_validation_method import DcvValidationMethod
+from open_mpic_core.mpic_coordinator.domain.mpic_orchestration_parameters import MpicRequestOrchestrationParameters
+from open_mpic_core.common_domain.check_parameters import CaaCheckParameters, DcvCheckParameters
 
 
 class BaseMpicRequest(BaseModel, ABC):
