+++ tests/unit/test_cohort_creator.py
@@ -7,8 +7,8 @@ import pytest
 import yaml
 from pydantic import TypeAdapter
 
-from aws_lambda_python.mpic_coordinator.cohort_creator import CohortCreator
-from aws_lambda_python.common_domain.remote_perspective import RemotePerspective
+from open_mpic_core.mpic_coordinator.cohort_creator import CohortCreator
+from open_mpic_core.common_domain.remote_perspective import RemotePerspective
 
 
 # noinspection PyMethodMayBeStatic
