+++ tests/unit/test_mpic_dcv_request.py
@@ -1,9 +1,9 @@
 import json
 import pydantic
 import pytest
-from aws_lambda_python.common_domain.enum.check_type import CheckType
-from aws_lambda_python.common_domain.enum.dcv_validation_method import DcvValidationMethod
-from aws_lambda_python.mpic_coordinator.domain.mpic_request import MpicDcvRequest
+from open_mpic_core.common_domain.enum.check_type import CheckType
+from open_mpic_core.common_domain.enum.dcv_validation_method import DcvValidationMethod
+from open_mpic_core.mpic_coordinator.domain.mpic_request import MpicDcvRequest
 
 from valid_mpic_request_creator import ValidMpicRequestCreator
