+++ tests/unit/test_mpic_request_validator.py
@@ -1,9 +1,9 @@
 import sys
 import pytest
 
-from aws_lambda_python.common_domain.enum.dcv_validation_method import DcvValidationMethod
-from aws_lambda_python.mpic_coordinator.messages.mpic_request_validation_messages import MpicRequestValidationMessages
-from aws_lambda_python.mpic_coordinator.mpic_request_validator import MpicRequestValidator
+from open_mpic_core.common_domain.enum.dcv_validation_method import DcvValidationMethod
+from open_mpic_core.mpic_coordinator.messages.mpic_request_validation_messages import MpicRequestValidationMessages
+from open_mpic_core.mpic_coordinator.mpic_request_validator import MpicRequestValidator
 from valid_mpic_request_creator import ValidMpicRequestCreator
